Rearrange Julia code a bit
Split src/gap2.jl into src/adapter.jl and src/packages.jl
