+++ src/help.jl
@@ -1,3 +1,5 @@
+## enable access to GAP help system from Julia
+
 function GAP_help_string(topic::String, onlyexact::Bool = false)
     return GAP.gap_to_julia(GAP.Globals.HelpString(GAP.julia_to_gap(topic), onlyexact))
 end
